+++ modules/dynamic-routing-gateway/main.tf
@@ -12,8 +12,8 @@ resource "oci_core_drg_attachment" "drg_attachments" {
   display_name       = each.key
   drg_route_table_id = lookup(each.value.optionals, "drg_route_table_id", oci_core_drg.drg.default_drg_route_tables[0].vcn)
   network_details {
-    id             = each.value.id
-    type           = each.value.type
+    id             = each.value.vcn_id
+    type           = "VCN"
     route_table_id = lookup(each.value.optionals, "route_table_id", null)
     vcn_route_type = lookup(each.value.optionals, "vcn_route_type", null)
   }