+++ modules/dynamic-routing-gateway/output.tf
@@ -1,0 +1,6 @@
+output "drg" {
+  value = oci_core_drg.drg
+}
+output "attachments" {
+  value = oci_core_drg_attachment.drg_attachments
+}