+++ releases.md
@@ -1,6 +1,7 @@
 # v2.13.0:
 ## **New**
 * Introduce `network-firewall` module to manage network firewalls and network firewall policies.
+* Introduce `dynamic-routing-gateway` module to manage dynamic routing gateway (DRG) and its attachments.
 * `network`: Add support for IPv6 via optional `ipv6` variable. 
   * Add new variables for IPv6 configuration:
     * `ipv6.enabled`